modified makefile and added libvmaf files
