+++ wrapper/src/libvmaf.cpp
@@ -1,0 +1,31 @@
+#include "vmaf.h"
+#include "libvmaf.h"
+#include <cstdio>
+
+double compute_vmaf(char* fmt, int width, int height, char *ref_path, char *dis_path)
+{
+    char *model_path;
+    char *log_path = NULL;
+    char *log_fmt = NULL;
+    bool disable_clip = false;
+    bool disable_avx = false;
+    bool enable_transform = false;
+    bool do_psnr = false;
+    bool do_ssim = false;
+    bool do_ms_ssim = false;
+    char *pool_method = NULL;
+
+
+    try
+    {
+        double score = RunVmaf(fmt, width, height, ref_path, dis_path, model_path, log_path, log_fmt, disable_clip, enable_transform, do_psnr, do_ssim, do_ms_ssim, pool_method);
+
+        return score;
+    }
+    catch (const std::exception &e)
+    {
+        fprintf(stderr, "Error: %s\n", e.what());
+        return -1;
+    }
+
+}
