+++ wrapper/src/libvmaf.h
@@ -1,0 +1,1 @@
+double compute_vmaf(char* fmt, int width, int height, char *ref_path, char *dis_path);
